refactor: use immutable objects
The current implementation uses a mutable data structure (ClusterData)
that is passed to several functions and modified. For developers not
familiar with code this can be difficult to trace since fields can be
set or modified many places in the code.

This change makes use of records and unmodifiable maps to ensure state
is not modified after creation.